Remove broken link target special handling
Symlinks can just be handled like normal files. Their size is the length
of the target string since the body is the target string. Because of
this it can easily just be written like normal files.

The special handling of having a dedicated field in the header struct blocks
normal usage since it is not actually written to the header in the archive.

This commit removes that special handling.